requirements.txt
Flask
twilio
requests
openai-whisper
torch
git+https://github.com/coqui-ai/TTS
